Optimize file structure - rename repo
