+++ cmd/interpreter_api/main.go
@@ -15,56 +15,17 @@ import (
 	"sync"
 	"time"
 
+	"github.com/NotCoffee418/european_smart_meter/pkg/types"
 	"github.com/gorilla/websocket"
 	"github.com/jacobsa/go-serial/serial"
 	"github.com/sigurn/crc16"
 )
 
-type MeterReading struct {
-	Timestamp string `json:"timestamp"`
-
-	// Current consumption/production
-	CurrentConsumptionKW float64 `json:"current_consumption_kw"`
-	CurrentProductionKW  float64 `json:"current_production_kw"`
-	L1ConsumptionKW      float64 `json:"l1_consumption_kw"`
-	L2ConsumptionKW      float64 `json:"l2_consumption_kw"`
-	L3ConsumptionKW      float64 `json:"l3_consumption_kw"`
-	L1ProductionKW       float64 `json:"l1_production_kw"`
-	L2ProductionKW       float64 `json:"l2_production_kw"`
-	L3ProductionKW       float64 `json:"l3_production_kw"`
-
-	// Totals
-	TotalConsumptionDayKWH   float64 `json:"total_consumption_day_kwh"`
-	TotalConsumptionNightKWH float64 `json:"total_consumption_night_kwh"`
-	TotalProductionDayKWH    float64 `json:"total_production_day_kwh"`
-	TotalProductionNightKWH  float64 `json:"total_production_night_kwh"`
-
-	// Electrical info
-	CurrentTariff int     `json:"current_tariff"`
-	L1VoltageV    float64 `json:"l1_voltage_v"`
-	L2VoltageV    float64 `json:"l2_voltage_v"`
-	L3VoltageV    float64 `json:"l3_voltage_v"`
-	L1CurrentA    float64 `json:"l1_current_a"`
-	L2CurrentA    float64 `json:"l2_current_a"`
-	L3CurrentA    float64 `json:"l3_current_a"`
-
-	// Switches/status
-	SwitchElectricity int `json:"switch_electricity"`
-	SwitchGas         int `json:"switch_gas"`
-
-	// Serial numbers
-	MeterSerialElectricity string `json:"meter_serial_electricity"`
-	MeterSerialGas         string `json:"meter_serial_gas"`
-
-	// Gas
-	GasConsumptionM3 float64 `json:"gas_consumption_m3"`
-}
-
 type P1Reader struct {
 	port           string
 	baudrate       uint
 	serialPort     io.ReadWriteCloser
-	latestReading  *MeterReading
+	latestReading  *types.MeterReading
 	readingMutex   sync.RWMutex
 	wsClients      map[*websocket.Conn]bool
 	wsClientsMutex sync.RWMutex
@@ -159,13 +120,13 @@ func (p *P1Reader) ValidateCRC(telegram string) bool {
 	return strings.ToUpper(givenCRC) == calcCRCHex
 }
 
-func (p *P1Reader) ParseTelegram(telegram string) *MeterReading {
+func (p *P1Reader) ParseTelegram(telegram string) *types.MeterReading {
 	if !p.ValidateCRC(telegram) {
 		log.Println("Invalid CRC, skipping telegram")
 		return nil
 	}
 
-	reading := &MeterReading{
+	reading := &types.MeterReading{
 		Timestamp: time.Now().Format(time.RFC3339),
 	}
 
@@ -308,22 +269,19 @@ func (p *P1Reader) StartReading() {
 
 			p.BroadcastToWebSockets(reading)
 			consecutiveErrors = 0
-			log.Printf("Successfully parsed telegram - Current consumption: %.3f kW", reading.CurrentConsumptionKW)
 		}
-
-		time.Sleep(100 * time.Millisecond)
 	}
 
 	log.Printf("Too many consecutive errors (%d), stopping reader", maxErrors)
 }
 
-func (p *P1Reader) GetLatestReading() *MeterReading {
+func (p *P1Reader) GetLatestReading() *types.MeterReading {
 	p.readingMutex.RLock()
 	defer p.readingMutex.RUnlock()
 	return p.latestReading
 }
 
-func (p *P1Reader) BroadcastToWebSockets(reading *MeterReading) {
+func (p *P1Reader) BroadcastToWebSockets(reading *types.MeterReading) {
 	p.wsClientsMutex.RLock()
 	clients := make([]*websocket.Conn, 0, len(p.wsClients))
 	for client := range p.wsClients {
